feat: add InventoryLineDetailsFixedItems components
